refactor: melhoria nas classes e adiciona configuração para UserDetails
