+++ eazybytes/src/main/java/com/springsecurity/eazybytes/customer/repository/CustomerRepository.java
@@ -1,5 +1,6 @@
-package com.springsecurity.eazybytes.customer;
+package com.springsecurity.eazybytes.customer.repository;
 
+import com.springsecurity.eazybytes.customer.entity.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
