+++ eazybytes/src/main/resources/application.yaml
@@ -17,10 +17,10 @@ spring:
     show-sql: false
     hibernate:
       ddl-auto: update
-  security:
-    user:
-      name: eazybytes
-      password: 123456
+#  security:
+#    user:
+#      name: eazybytes
+#      password: 123456
 
 
 # Logs
